fix: integrate revenue module into workspace contract structure
- Fix contracts/Cargo.toml: remove duplicate [package] section,
  make it a proper workspace root only
- Move revenue module from contracts/src/ to contracts/subscription/src/
  adapted for the new storage-contract delegation pattern
- Add revenue StorageKey variants to contracts/types/src/lib.rs:
  RevenueRecognitionRule, RevenueSchedule, RevenueDeferredBalance,
  RevenueRecognisedBalance, RevenueMerchantSubscriptions
- Hook revenue schedule generation into charge_subscription in
  contracts/subscription/src/lib.rs
- Add revenue public API to SubTrackrSubscription:
  set_revenue_rule, recognize_revenue, get_deferred_revenue,
  get_revenue_schedule
- Fix clippy::manual_div_ceil in revenue.rs (use .div_ceil())
- Add #[allow(clippy::too_many_arguments)] to subscription and proxy
  crates (pre-existing upstream issue with Soroban contractimpl macro)
- Remove obsolete contracts/src/ directory
- All workspace tests pass, cargo fmt --check clean, clippy -D warnings clean
