+++ contracts/Cargo.toml
@@ -1,20 +1,3 @@
-[package]
-name = "subtrackr-subscription"
-version = "0.2.0"
-edition = "2021"
-authors = ["SubTrackr Team"]
-description = "SubTrackr subscription implementation contract (Soroban)"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-soroban-sdk = "21.0.0"
-subtrackr-types = { path = "../types" }
-
-[dev-dependencies]
-soroban-sdk = { version = "21.0.0", features = ["testutils"] }
-arbitrary = { version = "1.3", features = ["derive"] }
 [workspace]
 resolver = "2"
 members = [
